@@ -1,12 +1,14 @@
 package com.example.supjain.bakingapp.Adapters;
 
-import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
+import androidx.databinding.ViewDataBinding;
+import androidx.recyclerview.widget.RecyclerView;
+
 import android.view.LayoutInflater;
 import android.view.View;
-import android.support.v7.widget.RecyclerView;
 import android.view.ViewGroup;
 
+import com.example.supjain.bakingapp.BR;
 import com.example.supjain.bakingapp.data.RecipeData;
 import com.example.supjain.bakingapp.databinding.RecipeNamesListItemBinding;
 
@@ -19,7 +21,6 @@ public class RecipeDataAdapter extends RecyclerView.Adapter<RecipeDataAdapter.Re
 
     private List<RecipeData> recipeDataList;
     private RecipeAdapterOnClickHandler clickHandler;
-    private Context context;
 
     public RecipeDataAdapter(RecipeAdapterOnClickHandler clickHandler) {
         this.clickHandler = clickHandler;
@@ -33,9 +34,8 @@ public class RecipeDataAdapter extends RecyclerView.Adapter<RecipeDataAdapter.Re
     @NonNull
     @Override
     public RecipeDataAdapterViewHolder onCreateViewHolder(@NonNull ViewGroup viewGroup, int i) {
-        this.context = viewGroup.getContext();
-        LayoutInflater layoutInflater = LayoutInflater.from(context);
-        RecipeNamesListItemBinding binding =
+        LayoutInflater layoutInflater = LayoutInflater.from(viewGroup.getContext());
+        ViewDataBinding binding =
                 RecipeNamesListItemBinding.inflate(layoutInflater, viewGroup, false);
         return new RecipeDataAdapterViewHolder(binding);
     }
@@ -61,16 +61,16 @@ public class RecipeDataAdapter extends RecyclerView.Adapter<RecipeDataAdapter.Re
     public class RecipeDataAdapterViewHolder extends RecyclerView.ViewHolder
             implements View.OnClickListener {
 
-        private final RecipeNamesListItemBinding binding;
+        private final ViewDataBinding binding;
 
-        public RecipeDataAdapterViewHolder(RecipeNamesListItemBinding binding) {
+        RecipeDataAdapterViewHolder(ViewDataBinding binding) {
             super(binding.getRoot());
             this.binding = binding;
             binding.getRoot().setOnClickListener(this);
         }
 
-        public void bind(RecipeData data) {
-            binding.setRecipe(data);
+        void bind(RecipeData data) {
+            binding.setVariable(BR.recipe, data);
             binding.executePendingBindings();
         }
 
