@@ -1,15 +1,16 @@
 package com.example.supjain.bakingapp;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import com.example.supjain.bakingapp.Adapters.RecipeStepsAdapter;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * This fragment inflates view for holding recipe ingredients list and steps list.
@@ -18,7 +19,8 @@ public class RecipeStepsFragment extends Fragment {
 
     private RecipeStepsAdapter recipeStepsAdapter;
 
-    public RecipeStepsFragment(){}
+    public RecipeStepsFragment() {
+    }
 
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
