@@ -1,20 +1,21 @@
 package com.example.supjain.bakingapp;
 
 import android.app.Application;
-import android.arch.lifecycle.AndroidViewModel;
-import android.arch.lifecycle.LiveData;
-import android.databinding.ObservableField;
 
 import com.example.supjain.bakingapp.DB.RecipeRepository;
 import com.example.supjain.bakingapp.data.RecipeData;
 
 import java.util.List;
 
+import androidx.databinding.ObservableField;
+import androidx.lifecycle.AndroidViewModel;
+import androidx.lifecycle.LiveData;
+
 public class RecipeViewModel extends AndroidViewModel {
 
-    private RecipeRepository repository;
     public LiveData<List<RecipeData>> recipeList;
     public ObservableField<String> errorMsg;
+    private RecipeRepository repository;
 
     public RecipeViewModel(Application application) {
         super(application);
@@ -22,11 +23,11 @@ public class RecipeViewModel extends AndroidViewModel {
         recipeList = repository.getRecipes();
     }
 
-    public void insertRecipes(List<RecipeData> recipeDataList) {
+    void insertRecipes(List<RecipeData> recipeDataList) {
         repository.insertRecipes(recipeDataList);
     }
 
-    public void setErrorMsg(String errorMsg) {
+    void setErrorMsg(String errorMsg) {
         this.errorMsg = new ObservableField<>(errorMsg);
     }
 }
