@@ -1,9 +1,7 @@
 package com.example.supjain.bakingapp.Util;
 
 import android.annotation.SuppressLint;
-import android.databinding.BindingAdapter;
 import android.graphics.drawable.Drawable;
-import android.support.v7.widget.RecyclerView;
 import android.text.TextUtils;
 import android.widget.ImageView;
 
@@ -12,6 +10,9 @@ import com.example.supjain.bakingapp.data.RecipeData;
 import com.squareup.picasso.Picasso;
 
 import java.util.List;
+
+import androidx.databinding.BindingAdapter;
+import androidx.recyclerview.widget.RecyclerView;
 
 public class BindingUtil {
 
@@ -26,9 +27,9 @@ public class BindingUtil {
 
     @SuppressLint("CrossDomainBindingAdapterIssue")
     @BindingAdapter({"data"})
-    public static void setRecyclerViewDataList(RecyclerView view, List<RecipeData> data)
-    {
+    public static void setRecyclerViewDataList(RecyclerView view, List<RecipeData> data) {
         RecipeDataAdapter adapter = (RecipeDataAdapter) view.getAdapter();
-        adapter.setRecipeData(data);
+        if (adapter != null)
+            adapter.setRecipeData(data);
     }
 }
