@@ -1,18 +1,17 @@
 package com.example.supjain.bakingapp.data;
 
-import android.arch.persistence.room.ColumnInfo;
-import android.arch.persistence.room.Entity;
-import android.arch.persistence.room.PrimaryKey;
-import android.arch.persistence.room.TypeConverters;
 import android.os.Parcel;
 import android.os.Parcelable;
-import android.support.annotation.NonNull;
 
-import com.example.supjain.bakingapp.RecipeStepsActivity;
 import com.google.gson.annotations.SerializedName;
 
-import java.util.ArrayList;
 import java.util.List;
+
+import androidx.annotation.NonNull;
+import androidx.room.ColumnInfo;
+import androidx.room.Entity;
+import androidx.room.PrimaryKey;
+import androidx.room.TypeConverters;
 
 /**
  * This class object holds all the recipe details.
@@ -21,28 +20,34 @@ import java.util.List;
 @TypeConverters(RecipeTypeConverters.class)
 public class RecipeData implements Parcelable {
 
+    public static final Creator<RecipeData> CREATOR = new Creator<RecipeData>() {
+        @Override
+        public RecipeData createFromParcel(Parcel in) {
+            return new RecipeData(in);
+        }
+
+        @Override
+        public RecipeData[] newArray(int size) {
+            return new RecipeData[size];
+        }
+    };
     @PrimaryKey
     @NonNull
     @ColumnInfo(name = "id")
     @SerializedName("id")
     private int recipeId;
-
     @ColumnInfo(name = "name")
     @SerializedName("name")
     private String recipeName;
-
     @ColumnInfo(name = "ingredients")
     @SerializedName("ingredients")
     private List<RecipeIngredientsData> recipeIngredients;
-
     @ColumnInfo(name = "steps")
     @SerializedName("steps")
     private List<RecipeStepsData> recipeSteps;
-
     @ColumnInfo(name = "servings")
     @SerializedName("servings")
     private int recipeServings;
-
     @ColumnInfo(name = "image_url")
     @SerializedName("image")
     private String recipeImageUrl;
@@ -57,28 +62,17 @@ public class RecipeData implements Parcelable {
         this.recipeImageUrl = recipeImageUrl;
     }
 
-    public void setRecipeId(@NonNull int recipeId) {
-        this.recipeId = recipeId;
+    protected RecipeData(Parcel in) {
+        recipeId = in.readInt();
+        recipeName = in.readString();
+        recipeIngredients = in.readArrayList(getClass().getClassLoader());
+        recipeSteps = in.readArrayList(getClass().getClassLoader());
+        recipeServings = in.readInt();
+        recipeImageUrl = in.readString();
     }
 
-    public void setRecipeName(String recipeName) {
-        this.recipeName = recipeName;
-    }
-
-    public void setRecipeIngredients(List<RecipeIngredientsData> recipeIngredients) {
-        this.recipeIngredients = recipeIngredients;
-    }
-
-    public void setRecipeSteps(List<RecipeStepsData> recipeSteps) {
-        this.recipeSteps = recipeSteps;
-    }
-
-    public void setRecipeServings(int recipeServings) {
-        this.recipeServings = recipeServings;
-    }
-
-    public void setRecipeImageUrl(String recipeImageUrl) {
-        this.recipeImageUrl = recipeImageUrl;
+    public static Creator<RecipeData> getCREATOR() {
+        return CREATOR;
     }
 
     @NonNull
@@ -86,28 +80,48 @@ public class RecipeData implements Parcelable {
         return recipeId;
     }
 
+    public void setRecipeId(@NonNull int recipeId) {
+        this.recipeId = recipeId;
+    }
+
     public String getRecipeName() {
         return recipeName;
+    }
+
+    public void setRecipeName(String recipeName) {
+        this.recipeName = recipeName;
     }
 
     public List<RecipeIngredientsData> getRecipeIngredients() {
         return recipeIngredients;
     }
 
+    public void setRecipeIngredients(List<RecipeIngredientsData> recipeIngredients) {
+        this.recipeIngredients = recipeIngredients;
+    }
+
     public List<RecipeStepsData> getRecipeSteps() {
         return recipeSteps;
+    }
+
+    public void setRecipeSteps(List<RecipeStepsData> recipeSteps) {
+        this.recipeSteps = recipeSteps;
     }
 
     public int getRecipeServings() {
         return recipeServings;
     }
 
+    public void setRecipeServings(int recipeServings) {
+        this.recipeServings = recipeServings;
+    }
+
     public String getRecipeImageUrl() {
         return recipeImageUrl;
     }
 
-    public static Creator<RecipeData> getCREATOR() {
-        return CREATOR;
+    public void setRecipeImageUrl(String recipeImageUrl) {
+        this.recipeImageUrl = recipeImageUrl;
     }
 
     @Override
@@ -124,25 +138,4 @@ public class RecipeData implements Parcelable {
         dest.writeInt(recipeServings);
         dest.writeString(recipeImageUrl);
     }
-
-    protected RecipeData(Parcel in) {
-        recipeId = in.readInt();
-        recipeName = in.readString();
-        recipeIngredients = in.readArrayList(getClass().getClassLoader());
-        recipeSteps = in.readArrayList(getClass().getClassLoader());
-        recipeServings = in.readInt();
-        recipeImageUrl = in.readString();
-    }
-
-    public static final Creator<RecipeData> CREATOR = new Creator<RecipeData>() {
-        @Override
-        public RecipeData createFromParcel(Parcel in) {
-            return new RecipeData(in);
-        }
-
-        @Override
-        public RecipeData[] newArray(int size) {
-            return new RecipeData[size];
-        }
-    };
 }
